units: Accept `u` as SI prefix for "micro-"
The `u` prefix, as in `10_us`, is not really standard, but presently
not ambiguous, since there is no real `u` prefix and no existing unit
that begins with `u` and could clash. The clash would be resolved in
favor of the existing unit anyway, just like `min` is not interpreted
as `milli-inch`.

Having this prefix makes it possible to enter all the SI prefixes in
trans-alpha mode.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>