simulator: Adjust minimum width for macOS
The windows look is all messed up on macOS with the previous change.
Trying with an intermediate value.
Note that if the screen aspect ratio is incorrectly detected, which is
the case over X11 and may be the case for WSL, you can use the `-s 2`
option when running on a Hi-DPI screen.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>