solver: Add compatibility mode for `MSLV`
The HP50G `MSLV` command, solving multiple equations with multiple
interdependent variables, is covered by `ROOT` on DB48x after the
introduction of the Jacobian solver. However, `MSLV` does not operate
as a function, but as a command leaving the equation and variables on
the stack. Replicate this behaviour for compatibility.

On DB48x, `ROOT` and `MSLV` are otherwise identical, with a single
underlying adaptative solving engine, selecting what takes 3 different
commands on HP50G (`ROOT`, `MROOT` and `MSLV`).

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>