equations: Fix sign error for buoyancy equation
In the example 2 of "Buoyancy & Terminal Velocity" ("Motion" section
of the equation library), we find the following solution:

```
[ vt=-1.14234 81034 5 m/s v=-0.79446 37698 68 m/s
  tfr=0.21337 88142 9 s xfr=0.15488 56277 51 m ]
```

where the first 3 values are the right ones (but not the fourth)
corresponding respectively to:

* `vt < 0` which is the expected negative value of the terminal
  velocity for an ascending bubble
* `v < 0` which is the expected negative speed achieved after `t=0.1 s`
* `tfr` the time required to reach the fraction `fr=0.95` of `vt`
* `xfr` the displacement during `tfr` which should be negative here
  corresponding to an ascending motion.

The problem comes from a probable error in the fourth equation which
is an integral of the hyperbolic tangent:

```
"'xfr_ft=(vt_(m/s))∫(0_s;(tfr_s)/(1_s);TANH(tⒸg/vt_(m/s));t)' "
```

where `vt` (which is negative) appears at 2 places. And since `TANH`
is an odd function, it appears clearly that the result will always be
positive. This means that one of the 2 `vt` should be in an absolute
value function. And it should be the one appearing in the `TANH`
argument because this is a time integral which should give a positive
duration:

```
Δt = ∫f(t) dt > 0

```

This should also be corrected in the reference
https://en.wikipedia.org/wiki/Terminal_velocity.

In our case this integral needs to be corrected accordingly in 2
places, since it appears also in the previous sub-section "Terminal
Velocity" as well as in "Buoyancy & Terminal Velocity", the equation
becoming:

```
'xfr_ft=(vt_(m/s))∫(0_s;(tfr_s)/(1_s);TANH(tⒸg/ABS(vt_(m/s)));t)'
```

Fixes: #1425

Originally titled Update equations_JW_25_03_09PM1.cc (code) and
Update equations_JW_25_03_09PM1.md (tests). Merged as one.

Signed-off-by: Wiljea <143950411+Wiljea@users.noreply.github.com>

With updates of commit message (lifted from #1425):

Reviewed-by: Christophe de Dinechin <christophe@dinechin.org>