compatibility: Add truth logic for integers.
When evaluating a logical operation on integers, such as `42 7 XOR`,
compatibility with HP calculators would require to return `False`.
However, the current implementation returns a bitwise result in that
case, i.e. `45`.

Document this behaviour and control it with a new flag,
`TruthLogicForIntegers`. Default is the current behaviour, which is
quite practical when dealing with binary numbers.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>