parser: Fix parsing of `==`
The `==` operator is the only one that has two non-names, but it also
has the problem that it may not be sorted correctly in the ID table
depending on the other IDs. When changing the IDs table to add
commands like `COL+`, the bisection would terminate on `ID_TestEQ`
first, which would match.

Since we are testing for `=` because of the first anomaly, we may as
well test for it directly. Not elegant, but less costly than
alternatives.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>