decimal: Do not read past end of number in `decimal::compare`
The `decimal::compare` function has an optional argument that can be
used to limit the comparision range. This is used in `sqrt` and
`cbrt`, and it could cause incorrect comparision results if the
numbers were trucated below that precision.

This impacted the computation of `sqrt(163)` at 24 digits of
precision. With either 25 or even 23 digits, the compuation would
correctly produce `12.76714 53348 03704 66171 1`. However, with 24
digits, the comparision would incorreclty terminate early due to the
bug, and `sqrt` would return `12.76714 53348 03704 66173 67152`.

This impacted the computation mentioned in #1440.

Fixes: #1440 (requires another fix for `ln` as well)

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>