decimal: Compute logarithms with extended precision
Computing `'ln(640320^3+744)'` with 24-digit precision would give an
incorrectly rounded result:

`40.10916 99911 32519 75535 07` instead of
`40.10916 99911 32519 75535 01`

This was due to a missing use of a `precision_adjust`, which needs to
be set in the outermost function and not just the inner ones. Compute
intermediate results with 3 additional digits of precision to ensure
accuracy to the last specified digit.

Fixes: #1440

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>