directory: Fix `STOVX` and `Store` to non-current directories
When writing a variable in a directory, the code adjusting the size of
enclosing directories only works if the directory being written into
is one of the current directories.

Ensure this condition by making it current.

Also fix a problem with a missing GC pointer for the symbol argument
in `polynomial::main_variable`, and a risk with the payload being read
from a `this` pointer that could theoretically move if the code is
changed.

Fixes: #1442

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>