decimal: Adjust tests to return to the values we should have
Return values for results that involve a square root to what they were
at 0.9.2, since this turns out to be better than what we had in 0.9.3.

This is notably true for Gamma function, which 0.9.2 got correctly
according to Wolfram Alpha, but 0.9.3 was computing incorrectly. The
reason is that this is the only function where we dynamically adjust
the number of additional precision digits. The adjustment is now
incorrectly blocked by the `precision_adjust::adjusted` not being `0`.

Also do the proper rounding for constant values. The last digit of `e`
for example should be a `9` and not a `8` at 24 digits. This is
addressed in a follow-up commit.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>