functions: Adjust unit conversion in `StdRnd`
The `StdRnd` command was just looking at the values of its arguments,
but after a conversion to base unit for the second one that would turn
`0.11_g` into `0.00011_kg` (the kg being the `UBASE` unit). This would
give incorrect results.

Instead, perform a conversion from second argument to the type of the
first one. This also checks unit consistency in a more solid manner.

Also added some unit tests for these specific cases and for error
cases (e.g. passing a unit to second argument of `RelRnd`).

Also added the commands to the `PartsMenu` and reorganized its pages
by topics.

Fixes: #1448

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>