power: Wait in `Off` command for key on
In order to be able to run a program that includes the `Off` command
correctly, we need to wait for the power state to be taken into
account by the firwmare. Also, since this may suspend the screen
refresh, run `lcd_refresh_wait()` first.

We do not draw the off image when `Off` is used programmatically.
It is better to display whatever the program had shown on screen.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>