stack: Fix display of size in interactive stack
On physical calculators, the `%zu` format for `printf` is no longer
supported after switching to the nano library. This caused the display
of information in the interactive stack to be wrong, showing `zu`
instread instead of the size, and then showing bogus information for
the type.

Replace `%zu` with `%lu` and convert argument to the `ulong` time.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>