graphics: Add `→HPGrob`, `→Bitmap`, `→Pixmap` and `Blank` commands
Add commands that generate or convert graphic objects:

* `→HPGrob` generates a graphic object in HP-compatible format.
* `→Bitmap generates a packed bitmap format (DB48x format)
* `→Pixmap`generates a color RPL pixmap format
* `Blank` generates a bitmap filled with the `Background` pattern

The `→HPGrob`, `→Bitmap` and `→Pixmap` commands render non-graphic
objects into graphic objects, using the `ResultFont`, `Foreground` and
`Background` settings. Text is rendered without the surrounding quotes.

If given a graphic object as input, they convert it to the target
format. On color RPL, pixmaps are dithered if converted to a GROB or
bitmap format.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>