sim: Correct hwfp entry on Mac with french locale.
On a Mac with french locale (at least), entry of hwfp (either
explicit with an `F` or `D` suffix or implicit via the `HwFP`
option) ignores the fractional part, because `strtof` and
`strtod` expect a `,` as decimal separator.

This also made the corresponding tests to fail.

Fix: reset the locale to "C" just before the RPL loop.

Signed-off-by: be-philippe <be.philippe@icloud.com>