general: bug with recent compiler (GCC >= 12.2.1)
With code optimized by a recent GCC compiler, the value of some `gcp`
pointer is kept in a CPU register, and the modification to that
same pointer through it's `operator+()` function provided to
`leb128<size_t>()` may be lost.

Explicitely returning the modified pointer from a `leb128<size_t>()`
with a different signature fixes the base bug.

A temporary hack is provided with a forced INLINE thunk with the
previous signature.

Fixes: #1458, #1286

Signed-off-by: be-philippe <be.philippe@icloud.com>