power: Avoid continuous draw on battery due to timers
It turns out that having timers active cause a continuous battery draw
of about 4.1mA, which lowers the displayed battery voltage by 0.3V and
shortens battery life.

Reorganize the code to use clock wakeups instead, which do not seem to
suffer the same high-current draw, allowing the calculator to use a
lower power mode, and only activate timers if there is less than one
second to go.

Thanks for HP Museum Forum user raprism for pointing me to the
discussion on the C47 and WP43 forums about similar problems on their
side. Apparently, everyone, not just me, was surprised that the timer
API would leave the CPU busy. I was fully expecting timers on such a
low-power system to be able to wakeup the CPU without drawing so much
current.

https://www.hpmuseum.org/forum/thread-20157-post-202670.html#pid202670

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>