expressions: Return errors for unknown primitives / derivatives
When we put an expression involving derivatives and primitives in an
expression, we need to make sure that the arguments are not evaluated
before we run the operation. For example, `'∂x(2·IP(3·x²))'` should
give the same error as `'2·IP(3·x²)' 'x' ∂`.

To achieve that objective, wrap the parameters of `Derivative` and
`Primitive` in expressions, so that after evaluation, we still get the
original expression and not its evaluated variant if applicable.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>