parser: Do not wrap variables after `Copy`
When parsing `'A+A▶A+A'`, we end up generating `'(A+A▶A)+A'`, but then
the output program was incorrectly quoting the second `A`. In other
words, `ToProgram` would give `« A A + 'A' ▶ 'A' + »` instead of the
expected `« A A + 'A' ▶ A + »`. This is because the `special` variable
remained set after having insert the `Copy` command.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>