time: Do not give truncated result for `FromHMS`
When converting `10.3033` using `FromHMS`, the result is inexact,
because the input is first converted to a fraction, and in that
specific case the precision of the fraction conversion is not
sufficient to give us the correct result. This is caused by the
default value of `FractionIterations` being a tad bit too small.

Thiw manifested with an even simpler test:

```rpl
10.3033 ToFraction ToDecimal
```

That would not give `10.3033` back, because the intermediate fraction
was not `103 033/10 000` but `19669/1909`, which does not evaluate
back to the original value.

Fixes: #1480

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>
Reported-by: User mateo222 on HPMuseum forum