rewrites: Correctly rewrite down for `rewrites<DOWN>`
We were passing the `down` flag from the `down` argument, which is an
enum with value `0` for `DOWN`. As a result, the `down` flag was
incorrectly cleared.

This led to really bad performance for primitive and derivative
operations, which work really much better going down than up, since
otherwise we re-evaluate the same thing over and over again.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>