expressions: Correctly substitute names in list substitutions
Substitutions described by a list were failing when the object being
substituted was a name.

The reported failing test case was

```rpl
'(A-2+sin(6*C))^J'
{A V J 9}
WHERE
```

Two `if` were inverted, causing the `V` to be taken as a name for the
next substitution instead of being taken as a substituted value.

Reported-by: MOR330z
Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>