solver: Do not emit type error when guesses are constants
The following test case was incorrectly reporting a type error when
solving for the `m` variable:

```rpl
m=Ⓒme
'MSlv(ⒺRelativityMassEnergy;[E];[1 eV])'
Eval
Pick3
StEq
SolvingMenu
```

The problem was that the initial guess for `m` was a constant object,
and that testing whether it was zero would cause the type error.

This is fixed by passing `false` to `is_zero` to avoid errors.

Reported-by: User raprism on HP Museum forum
Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>