array: Add support for power
Add support for power operations with arrays and vectors.

When given an integer value on the right, we perform matrix operations
as was the case before. Otherwise, we perform element-wise operations.
This means that `[[a b][c d] 2 ^` and `[[a b][c d] 2. ^` perform
different actions. This is intentional.

Also note that with vectors, this may generates higher-order vectors
or matrices. For example, `[a b] [c d] ^` generates a 2x2 matrix.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>