decimal: Show integers and fractions as decimal
A new setting, `ShowAsDecimal`, ensures that integers and fractions
show with decimal rendering when shown on the stack. This makes it
possible to benefit from formatting options such as `FIX` or to use
scientific notation for large integers.

Note that this does not apply to the `Show` command, where the
underlying object shows normally.

Fixes: #1498

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>