files: Fix saving of misaligned BMP files
The BMP-writing code was designed with HP-style GROB memory layout,
and had been broken when we started generating packed bitmaps.
As a quick workaround, convert the input bitmaps to GROB format before
running the existing code.

It's likely the code can be rewritten to take the new stride into
account, but the way to do that is not entirely obvious.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>