constants: Accelerate / improve the display of constant values
The display of constant values was displaying the expression used to
compute the constant's numerical value. This was unreadable. Add a
fourth field in the array used to define constants, where the fourth
entry is the approximate value to be displayed. This avoids expensive
computations just to build the menu.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>