tests: Cannot use the `ClearThingsMenu` while typing
The ClearThings menu is not available while typing, so the logic to
select it picks up Shift-Backspace, which deletes forward and beeps
instead of actually entering the menu. This manifests strangely as
curious errors about the shift state not matching expectation, simply
because the beep caused a delay in execution...

Spell out the command to avoid the problem.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>