variables: Various improvements to `Purge` and `PurgeAll`
The `PurgeAll` command would only accept names. It is really
practically if it also accepts whatever `Purge` accepts,
notably settings.

When the current directory is gone, report it.

When an error happens, e.g. purging a local name, report it.

Actually use the `directory::purge_all` function, which was not used
by the `PurgeAll` command implementation.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>