range: Fix the implementation of trig functions
The range implementation for trig functions had several bugs,
all related to the detection of the quadrant for the input values.

1. The function used to get the input angle as a multiple of pi
   was computing it modulo the input circle, which is undesirable when
   the goal is to estimate how many cycles of the functions we cross.

2. The computation of the quadrant was not continuous between positive
   and negative values, because of how the the C `%` operator changes

Some effort was made to also make the code more readable, with
variable name changes and adding a helper function to make the goal of
some computations easier to understand.

Added testing for the various cases.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>