constants: Add commands for pi, e, infinity and negative infinity
On HP calculators, it is possible to use `pi` as a command to get the
value of π. Users have reported that they were surprised the same was
not true on DB48x.

Add `pi` or `π` for the π constant.
Add `EulerianNumber` or `℮` for the e constant.
Add `Infinity` for a positive infinity.
Add `NegativeInfinity` for a negative infinity.

All these commands obey the `NumericalConstants` flag.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>