ranges: Put σ first when rendering uncertain numbers
Use the notation `1±σ3` instead of `1±3σ` for uncertain numbers.
Uncertain numbers are used for quadratic superposition.
The value shown on the right-hand side is the value of the standard
deviation σ for that number.
The notation `1±3σ` incorrectly suggest that we have 3 times the
standard deviation, and that's how it is often used.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>