src: Allow for "Julian day number"s with a negative Gregorian year
The function julian_day_number(...) in datetime.cc handled a calculation with the month incorrectly which leads to a wrong "Julian day number". The intermediate value rm is now an int (i.s.o. uint). Months are now handled correctly.

The function date_from_julian_day(...) in datetime.cc did not allow for a negative year which leads to a wrong year for smaller "Julian day number"s. Year is now an int (i.s.o. uint). The sign of a negative year is preserved and returned.

Added comments to both functions about the (most probable) source (references with URLs) of the algorithms used and their working. That took some reverse engineering. Pitty the original author didn't include more information.

Fixes: #1538

Other functions with dates may not handle negative years correctly. That might be(come) an issue but not part of this one.

In test.cc added 2 tests for conversion between "Julian day number" and (Gregorian) date.

Signed-off-by: Ed@vanGasteren.net