src: Allow for "Julian day number"s with a negative Gregorian year
algebraic_p date_from_julian_day(object_p jdn, bool error)
dit not work for smaller "Julian day number"s which should give a negative Gregorian year.
The year variable was a uint. It is now an int.
The sign of a negative year is preserved and returned.

Fixes: #1538

ularge julian_day_number(int d, int m, int y)
handled a calculation with the month incorrectly which could lead to a wrong "Julian day number".
The intermediate variable rm was a uint. It is now an int.

Added comments to both functions about the (most probable) source (references with URLs) of the algorithms used and their working.

This solves the reported issue.
But other functions with dates may not handle negative years correctly.
That might be(come) an (other) issue.

Signed-off-by: Ed@vanGasteren.net