src: make radian the base unit for angles
The base unit command for any angle unit returns a turn unit.
But the SI unit for a two-dimensional angle is the radian (symbol: rad).
The unit radian is available with the symbol: r.
The symbol rad is in use for a radiation unit.
To make radian the base unit, for now keeping symbol r,
its definition is swapped with the definition of turn.

Fixes: #1536
Signed-off-by: Ed@vanGasteren.net