src: move Temperature and Pressure units to separate (sub)menus
The units for Pressure where hard to find.
They where hidden in the Fluids (flow) menu.
That contained Temperature units as well.

Moved both the Pressere and Temperature units to their own menu.

That cauese several tests to fail because the units (sub)menus get rearranged.
The failing tests are adjusted (and commented).

Fixes: #1529.

A Signed-off-by: Ed@vanGasteren.net