src: command ToUncertain now leaves arguments unsorted
The range commands sort their arguments, since the two values
represent a lower and upper bound.

That should not be done for uncertain numbers, since the two values
are a mean and a standard deviation.

Included a test to verify that we do not do it.

Review: Reworded comment and commit message.

Fixes: #1515

Signed-off-by: Ed@vanGasteren.net
Reviewed-by: Christophe de Dinechin <christophe@dinechin.org>