runtime: Clear `ErrSrcLen` when clearing error
When we `end_edit` and there is an error, having a non-zero
`ErrSrcLen` can cause the `select` variable to be set. This, in turn,
changes the behaviour of the backspace key.

This was causing the `INPUT` command tests to fail when executed after
commands that were leaving an error source length in the
runtime. Specifically, running the `TruthPlot` complex example was
causing unexpected failures in the following `Input` tests.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>