src: fix Date+ crash in case of illegal date
The Date+ command, when given an illegal date, only partly recovers from the error.
It proceeds to the same operation with the arguments reversed and then crashes.

The functionality of reversed arguments is not documented.

Removed the code for reversed arguments.
Then the recovery of an earlier error gets completed.

Fixes: #1556
Signed-off-by: Ed@vanGasteren.net