ui: Further alpha/trans-alpha mode clean-up
Always reset `lowercase` when resetting `alpha`.
This ensures that the "usr/1us" annuciator values are never shown
and `lowercase` is always clear when entering non-trans-alpha mode.

Never reset `lowercase` when setting `alpha`.
This ensures that when entering trans-alpha mode the case is set
according to the trans-alpha key.

Restore alpha mode after using trans-alpha mode only when alpha mode
was not exited.

Signed-off-by: Thomas Eberhardt <sneakywumpus@gmail.com>
Reviewed-by: Christophe de Dinechin <christophe@dinechin.org>
Tested-by: Christophe de Dinechin <christophe@dinechin.org>