ios: Fix layout on physical machines
Physical devices like the iPhone 16 seem to be zooming the app a
little while running, losing some pixels on the edges. If you select
the app switcher, then the missing mixels show up, so it's
intentional on Apple's part, and not a problem with the layout being
wrong.

Make the left and right margins relative to the margin fixes that.
This makes the application look OK on iOS 18.5 for the iPhone 16 Pro.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>