ranges: Accept "natural" input values for creation function
When using `4 6 →∆Range`, we are expecting `4±6` and not `5±1`.
The range creation functions used to always take a low and high as
input values. They now take what they display as input.

In addition, the range create functions `→Range`, `→%Range`, `→∆Range`
and `→σRange` now accept a range as input and convert to that format.
It is not possible to convert between uncertain numbers and other
kinds of ranges.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>