android: Workaround display refresh bug
Whenever the screen has a non-zero offset, there is a display refresh
bug that shows up on the simulator, where the leftmost pixels of the
screen are refreshed with a delay. This seems to be a compositor bug,
but the only workaround I found after half a day of trying was to have
a larger pixmap that covers the whole screen width.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>