simulator: Reverse file installation logic
During early development, it was important to ensure that the
configuration files were updated every single time. As the project
matures, preserving installation files by default is probably a safer
option. Ensure that we run the simulator in the correct location,
which is important in particular when running on macOS from the
Finder.

Replace the `-c` option, which was hte the non-untuitive "nocopy"
option with a '-I' option which is the active "install" option and
resets files to what is bundled with the app. We also install if the
target directory is not found.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>