src: comply with SOFA Copyright and license
Before I inserted SOFA source code into src/datetime.*
because I could not get make to pick up additional files.
I found, at least for the simulator, a workaround,
by changing sim/db48x.pro.

Note: it isn't fixed for other builds.
It should be solved by adding the files to the Makefile
but I could not get that to work.

This commit adds:
- src/cal2jd.c
- src/jd2cal.c
- src/sofa.h
- src/sofam.h

They are "intact and unchanged copies of SOFA source code files"
not "derived work".

The previously SOFA source code inserted in src/datetime.* is removed.
Only a #include of the SOFA header files is inserted.

With this approach I believe to comply with SOFA Copyright and license.

ToDo: fix the Makefile or/and other build files.
At the moment that is beyond my competence.
@c3d please look into this.

Signed-off-by: Ed@vanGasteren.net